@@ -11,8 +11,8 @@
         }
     ],
     "require": {
-        "php": ">=7.3",
-        "guzzlehttp/guzzle": "^6.5.5|^7.0",
+        "php": ">=8.0",
+        "guzzlehttp/guzzle": "^7.5",
         "illuminate/config": "^7.0|^8.0|^9.0",
         "illuminate/support": "^7.0|^8.0|^9.0",
         "nesbot/carbon": "^2.17"
