@@ -3,53 +3,89 @@
 
 namespace Flooris\FileMakerDataApi\HttpClient;
 
-use Flooris\FileMakerDataApi\Client;
+use GuzzleHttp\Exception\GuzzleException;
+use Illuminate\Contracts\Cache\Repository as CacheRepositoryInterface;
+use Psr\Http\Message\ResponseInterface;
+use Psr\SimpleCache\InvalidArgumentException;
+use Flooris\FileMakerDataApi\Client as FmClient;
 use GuzzleHttp\RequestOptions;
+use GuzzleHttp\Client;
 
 class Connector
 {
-    /** @var string */
-    private $configHost;
+    private ?string $baseUrl;
+    private ?Client $guzzleClient = null;
 
-    private $guzzleClient;
-
-    public function __construct($configHost)
+    public function __construct(
+        private string                     $configHost,
+        protected CacheRepositoryInterface $cache
+    )
     {
-        $this->configHost   = $configHost;
-        $this->guzzleClient = new \GuzzleHttp\Client([
-            'base_uri' => $this->getBaseUri(),
+        $this->baseUrl = $this->getBaseUri();
+
+        if ($this->baseUrl === null) {
+            return;
+        }
+
+        $this->guzzleClient = new Client([
+            'base_uri' => $this->baseUrl,
         ]);
     }
 
-    public function get($uri, $query = [])
+    public function get(string $uri, ?string $sessionToken = null, array $query = []): ResponseInterface
     {
-        return $this->send('GET', $uri, null, $query);
+        return $this->send('GET', $uri, $sessionToken, null, $query);
     }
 
-    public function post($uri, $bodyData)
+    public function post(string $uri, ?string $sessionToken = null, mixed $bodyData = null): ResponseInterface
     {
-        return $this->send('POST', $uri, $bodyData);
+        return $this->send('POST', $uri, $sessionToken, $bodyData);
     }
 
-    public function patch($uri, $bodyData)
+    public function patch(string $uri, ?string $sessionToken = null, mixed $bodyData = null): ResponseInterface
     {
-        return $this->send('PATCH', $uri, $bodyData);
+        return $this->send('PATCH', $uri, $sessionToken, $bodyData);
     }
 
-    public function delete($uri)
+    public function delete(string $uri, ?string $sessionToken = null): ResponseInterface
     {
-        return $this->send('DELETE', $uri);
+        return $this->send('DELETE', $uri, $sessionToken);
+    }
+
+    public function getDataContainerToken(string $dataContainerObjectUrl): ?string
+    {
+        $options = [
+            RequestOptions::ALLOW_REDIRECTS => false,
+        ];
+
+        try {
+            $client = new Client([
+                'base_uri' => $dataContainerObjectUrl,
+            ]);
+
+            $response     = $client->request('GET', '', $options);
+            $cookieHeader = $response->getHeader('Set-Cookie');
+
+            if ($sessionToken = reset($cookieHeader)) {
+                return $sessionToken;
+            }
+
+            return null;
+
+        } catch (GuzzleException $e) {
+            return null;
+        }
     }
 
 
-    private function send($method, $uri, $bodyData = null, $query = [])
+    private function send(string $method, string $uri, ?string $sessionToken = null, mixed $bodyData = null, array $query = []): ResponseInterface
     {
         $options = [
             RequestOptions::HEADERS => [
                 'Content-Type'  => 'application/json',
                 'Accept'        => 'application/json',
-                'Authorization' => $this->getAuthorizationHeaderValue(),
-                'User-Agent'    => Client::USER_AGENT,
+                'Authorization' => $this->getAuthorizationHeaderValue($sessionToken),
+                'User-Agent'    => FmClient::USER_AGENT,
             ],
             RequestOptions::QUERY   => $query,
         ];
@@ -61,43 +97,40 @@ class Connector
         return $this->guzzleClient->request($method, $uri, $options);
     }
 
-    private function getBaseUri()
+    private function getBaseUri(): ?string
     {
-        $port = config(sprintf('filemaker.%s.port', $this->configHost));
+        $port     = config(sprintf('filemaker.%s.port', $this->configHost));
+        $protocol = config(sprintf('filemaker.%s.protocol', $this->configHost));
+        $host     = config(sprintf('filemaker.%s.hostname', $this->configHost));
+
+        if (! $protocol ||
+            ! $host
+        ) {
+            return null;
+        }
 
         return sprintf('%s%s%s/',
-            config(sprintf('filemaker.%s.protocol', $this->configHost)),
-            config(sprintf('filemaker.%s.hostname', $this->configHost)),
+            $protocol,
+            $host,
             $port ? ":{$port}" : '',
         );
     }
 
-    private function getAuthorizationHeaderValue()
+    private function getAuthorizationHeaderValue(?string $sessionToken = null): string
     {
-        $cacheKey = sprintf('filemaker.%s.session_token', $this->configHost);
-        if (cache()->has($cacheKey)) {
-            return 'Bearer ' . cache($cacheKey);
+        try {
+            if ($sessionToken) {
+                return "Bearer {$sessionToken}";
+            }
+
+            $username = config(sprintf('filemaker.%s.username', $this->configHost));
+            $password = config(sprintf('filemaker.%s.password', $this->configHost));
+
+            return 'Basic ' . base64_encode("{$username}:{$password}");
+        } catch (InvalidArgumentException $e) {
+            // ToDo: handle exception
         }
 
-        $usernameAndPassword = config(sprintf('filemaker.%s.username', $this->configHost)) . ':' .
-                               config(sprintf('filemaker.%s.password', $this->configHost));
-
-        return 'Basic ' . base64_encode($usernameAndPassword);
-    }
-
-    /**
-     * @return string
-     */
-    public function getSessionToken(): string
-    {
-        return $this->session_token;
-    }
-
-    /**
-     * @param string $session_token
-     */
-    public function setSessionToken(string $session_token): void
-    {
-        $this->session_token = $session_token;
+        return '';
     }
 }
